requirements.txt
PyQt5
mysql-connector-python
bcrypt
smtplib
email
random
datetime
